Updates to AsciiDoc parser
